restructures project like flask tutorial
