+++ flaskr/data/SPTickerWiki.csv
