+++ flaskr/scripts/analyses/runstatsclose.py
