+++ flaskr/static/IMG/advdecplt2019.png
