+++ flaskr/static/css/about.css
@@ -1,0 +1,7 @@
+body{
+  height: 100%;
+}
+.about {
+  background-color: lightgreen;
+  color: black;
+}
