+++ flaskr/static/css/home.css
@@ -1,0 +1,7 @@
+body{
+  height: 100%;
+}
+.main {
+  background-color: lightgreen;
+  color: black;
+}
