+++ flaskr/static/css/template.css
@@ -1,0 +1,12 @@
+body{
+  height: 100%;
+}
+
+.main {
+  background-color: beige;
+  color: black;
+} 
+.about {
+  background-color: lightgreen;
+  color: black;
+}
