+++ flaskr/templates/about.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    <meta charset="utf-8">
+    <title>About Flask</title>
+  </head>
+  <body>
+    {% extends "template.html" %}
+    {% block content %}
+    <div class="about">
+      
+      <h1> About Flask </h1>
+      <p> Flask is a micro web framework written in Python.</p>
+      <p> Applications that use the Flask framework include Pinterest,
+        LinkedIn, and the community web page for Flask itself.</p>
+    </div>
+      
+    {% endblock %}
+  </body>
+</html>
