+++ flaskr/templates/template.html
@@ -10,7 +10,6 @@
   <body>
     <header>
       <div class="container">
-        <h1 class="logo">First Web App</h1>
         <strong><nav>
           <ul class="menu">
             <li><a href="{{ url_for('home') }}">Home</a></li>
