--- virtual/bin/python3
@@ -1,1 +1,0 @@
-python3.9
