--- virtual/bin/python3.9
@@ -1,1 +1,0 @@
-/opt/homebrew/opt/python@3.9/bin/python3.9
