--- virtual/pyvenv.cfg
@@ -1,3 +1,0 @@
-home = /opt/homebrew/opt/python@3.9/bin
-include-system-site-packages = true
-version = 3.9.17
